// bgr-ui/src/app/services/models/page-response-board-game-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BoardGameResponse } from '../models/board-game-response';
export interface PageResponseBoardGameResponse {
  content?: Array<BoardGameResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}